Split CalendarController
Split into:
- CalendarController
- CalendarEventController
- CalendarImageController
